requirements.txt
Django==5.2.2
python-dotenv==1.0.1
django-allauth==0.61.1
whitenoise==6.6.0
gunicorn==21.2.0
# API and Authentication
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
# Database
#psycopg2-binary==2.9.9
dj-database-url==2.1.0
# API Documentation
drf-yasg==1.21.7
# Real-time features
channels==4.0.0
channels-redis==4.2.0
# Email and Notifications
django-celery-beat
celery==5.3.6
redis==5.0.1
# File handling
django-storages==1.14.2
#Pillow==10.2.0
# Security
django-ratelimit==4.1.0
django-axes==6.3.0
# Search
django-elasticsearch-dsl==7.3.0
# Testing
pytest==8.0.0
pytest-django==4.8.0
coverage==7.4.1
# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
firebase-admin==6.2.0
# Forms
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
requests 